requirements_agentic_rag.txt
streamlit>=1.32
altair<5
pandas
numpy
scikit-learn
joblib
langgraph>=0.2.50
faiss-cpu; python_version < "3.14"
